+++ app/package.json
@@ -61,11 +61,11 @@
     "@formatjs/intl-relativetimeformat": "9.3.1",
     "@hyperledger/anoncreds-react-native": "0.2.2",
     "@hyperledger/aries-askar-react-native": "0.2.3",
-    "@hyperledger/aries-bifold-core": "1.0.0-alpha.285",
-    "@hyperledger/aries-bifold-remote-logs": "1.0.0-alpha.285",
-    "@hyperledger/aries-bifold-verifier": "1.0.0-alpha.285",
-    "@hyperledger/aries-oca": "1.0.0-alpha.285",
-    "@hyperledger/aries-react-native-attestation": "1.0.0-alpha.285",
+    "@hyperledger/aries-bifold-core": "1.0.0-alpha.291",
+    "@hyperledger/aries-bifold-remote-logs": "1.0.0-alpha.291",
+    "@hyperledger/aries-bifold-verifier": "1.0.0-alpha.291",
+    "@hyperledger/aries-oca": "1.0.0-alpha.291",
+    "@hyperledger/aries-react-native-attestation": "1.0.0-alpha.291",
     "@hyperledger/indy-vdr-react-native": "0.2.2",
     "@hyperledger/indy-vdr-shared": "0.2.2",
     "@react-native-async-storage/async-storage": "1.15.11",
